fix: use new &tab= URL format
Since https://github.com/sourcegraph/sourcegraph/pull/12109, URLs with a view state (eg `references`) have used `#L123&tab=references` not `#L123$references`.

This commit fixes the URLs for the hover actions, which eliminates a `<Redirect>` (which, in turn, causes a full unmount-mount cycle for `BlobPage` and all of its children, which is slow).