feat: add optional position adjuster to HoverifierOptions (#38)
This commit adds the option to adjust positions used in codeintellify. This is needed because some code hosts change whitespace to fit their needs in the UI. In order to get correct code intelligence we need to be able to map the position that we found in the DOM to the actual position of that token in it's source.