fix: delay highlighting the hovered token until the hover is shown
Fixes an issue where the token would be highlighted before the hover was shown (if the hover was taking a while to load). The reason behind not showing the hover loading immediately is to avoid UI jitter. In keeping with this principle, it makes sense to not show any changes in the UI (including highlighting the hovered range) until we know they will not need to be immediately redrawn. If the token is highlighted and then a few hundred msec later the hover is determined to be empty, the highlight will go away. That looks jittery, and that problem is fixed by this commit.