feat: remove close button
The close button had problems:

- It was a small click target
- It obscured hover content or hover actions in some cases
- It especially looked weird when there were only actions and no hover content; it would obscure part of the rightmost button, not taking up the full width

The problems it solved were:

1. Informing users how to close the overlay
2. Indicating whether the overlay was pinned

For (1), I think users would figure it out that they need to click elsewhere or press the Esc key to close the overlay.

For (2), I don't think it was noticeable enough to effectively do this, and I don't think this is an important goal.