fix: overlay remains visible on click when already moused over
If the overlay is visible due to a mouseover event and the user clicks a token to pin it, previously the overlay would briefly disappear as its data was reloaded. This could be worked around in the consumer by having replayable observables, but most consumers did not implement this correctly. Now, when a click event pins the overlay (and does not change its position from what the prior mouseover event yielded), it does not unsubscribe the previous data observables; it remains visible and subscribed.